Update table creation scripts
The Players table now mirrors the stats given on basketball-reference.com. Primary key is not just the name and year of the player since the player name will not be uniquely identifiable. Additionally, the player names would take too much time to format as FIRST, LAST so we will keep it whole which will include middle names (hopefully reducing data conflicts).